itfMesManager: FUN_003a6c10 remove premature (short) cast on int array element - was triggering a narrow lh load instead of retail's full lw, a real correctness-adjacent codegen bug; diff 63->62, register-coloring floor remains on loop var

+++ src/itfMesManager.c
@@ -4632,7 +4632,7 @@ short * FUN_003a6c10(int param_1,short *param_2)
 
   for (; -1 < iVar2; iVar2 = iVar2 + -1) {
 
-    *param_2 = ((short)aiStack_30[iVar2] + 0x10) * 0x100 + -0x7f80;
+    *param_2 = (aiStack_30[iVar2] + 0x10) * 0x100 + -0x7f80;
 
     param_2 = param_2 + 1;
